Split bricks/__init__.py
