+++ examples/groundhog/language.py
@@ -12,7 +12,8 @@ from groundhog.mainLoop import MainLoop
 from groundhog.datasets import LMIterator
 from groundhog.trainer.SGD import SGD
 
-from blocks.bricks import VariableRole, Tanh
+from blocks.bricks import Tanh
+from blocks.bricks.base import VariableRole
 from blocks.bricks.recurrent import GatedRecurrent
 from blocks.select import Selector
 from blocks.graph import ComputationGraph
